requirements.txt
mcp-python
aiohttp
